Adding warning dialog for deleting music folders, redisigned adding music button as  list item.
