+++ ViewModel/MainViewModel.cs
@@ -8,19 +8,23 @@ namespace Kova.ViewModel
 {
     public class MainViewModel : ViewModelBase
     {
+        private readonly IDialogCoordinator _dialogCoordinator;
         private ViewModelBase _currentViewModel;
         private ViewModelLocator _VMLocator;
 
         public RelayCommand LaunchKovaCommand { get; private set; }
         public RelayCommand<ViewModelBase> ChangeViewCommand { get; private set; }
-       
+        public RelayCommand ShowMessegeDialogCommand { get; private set; }
+
         public MainViewModel()
         {
+            _dialogCoordinator = DialogCoordinator.Instance;
             _VMLocator = new ViewModelLocator();
             CurrentViewModel = _VMLocator.Player;
 
             ChangeViewCommand = new RelayCommand<ViewModelBase>((View) => ChangeView(View));
             LaunchKovaCommand = new RelayCommand(LaunchKova);
+            ShowMessegeDialogCommand = new RelayCommand(ShowDialogAsync);
         }
 
         public ViewModelBase CurrentViewModel
@@ -35,7 +39,19 @@ namespace Kova.ViewModel
                 RaisePropertyChanged(nameof(CurrentViewModel));
             }
         }
-     
+
+        private async void ShowDialogAsync()
+        {
+            var customDialog = new CustomDialog() { Title = "Add music" };
+            var addMusicDialog = new AddMusicDialogViewModel(instance =>
+            {
+                _dialogCoordinator.HideMetroDialogAsync(this, customDialog);
+            }, _dialogCoordinator);
+
+            customDialog.Content = new Views.AddMusicDialog() { DataContext = addMusicDialog };
+            await _dialogCoordinator.ShowMetroDialogAsync(this, customDialog);
+        }
+
         private void ChangeView(ViewModelBase other)
         {
             CurrentViewModel = other;
