+++ bin/Debug/Kova.vshost.exe.config
@@ -16,12 +16,6 @@
             <setting name="SelectedAccentIndex" serializeAs="String">
                 <value>0</value>
             </setting>
-            <setting name="MusicFolderPath" serializeAs="Xml">
-                <value>
-                    <ArrayOfString xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                        xmlns:xsd="http://www.w3.org/2001/XMLSchema" />
-                </value>
-            </setting>
         </Kova.Properties.Settings>
     </userSettings>
 </configuration>
